feat: web streams based encoding/csv
