+++ encoding/testdata/complex.csv
@@ -1,0 +1,7 @@
+id,name,email
+1,deno,deno@example.com
+# This is a comment
+
+2,node,"node@example.com"
+
+3,,"test@example.com"
