+++ encoding/testdata/simple.csv
@@ -1,0 +1,3 @@
+id,name
+1,foobar
+2,barbaz
